pyproject.toml
[project]
name = "google-image-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "bs4>=0.0.2",
    "google-api-python-client>=2.154.0",
    "gspread>=6.1.4",
    "openai>=1.54.4",
    "pillow>=11.0.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "selenium>=4.26.1",
]

[tool.uv]
dev-dependencies = [
    "black>=24.10.0",
]
